requirements.txt
flask
PyYAML
celery[redis]
